scraper/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
supabase>=2.0.0
pydantic>=2.0.0
tenacity>=8.2.0
httpx>=0.25.0
google-generativeai>=0.8.0
